Generalize BaseContextData.java
