+++ api/src/main/java/org/arend/ext/typechecking/BaseContextData.java
@@ -1,0 +1,22 @@
+package org.arend.ext.typechecking;
+
+import org.arend.ext.concrete.ConcreteSourceNode;
+import org.arend.ext.core.expr.CoreExpression;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public interface BaseContextData {
+  /**
+   * Returns the expected type in the current context.
+   * The meta definition should return an expression of this type.
+   * It can be {@code null} if the expected type is unknown.
+   */
+  CoreExpression getExpectedType();
+
+  void setExpectedType(@Nullable CoreExpression expectedType);
+
+  /**
+   * A marker that can be used for error reporting.
+   */
+  @NotNull ConcreteSourceNode getMarker();
+}
